requirements.txt
dash==2.18.1
pandas==2.2.2
plotly==5.24.1
statsmodels==0.14.2
snowflake-connector-python==3.12.2
gunicorn==23.0.0
